ims/requirements.txt
django
pillow
pytesseract
requests